Changed structs back to classes
This code was ported from Java and some functions still have
Java-styled functions that modify one or more class properties without
having ’out’ or ‘ref’ keywords on parameter list, breaking the logic
when classes are defined as value type instead of reference type as in
java. I changed these classes back to class from struct fixing those
issues and making sure that engine works as expected. That makes
particle simulation work as expected but that is not 100% correct way
to do this.